Created Generic Test Package
Added excluded in Ant compile to avoid test packages.
