+++ src/test/net/sf/memoranda/ReminderLogTest.java
@@ -1,4 +1,4 @@
-package net.sf.memoranda;
+package test.net.sf.memoranda;
 
 import static org.junit.Assert.*;
 
@@ -6,7 +6,9 @@ import org.junit.AfterClass;
 import org.junit.BeforeClass;
 import org.junit.Test;
 
-public class EventsLogTest {
+import net.sf.memoranda.ReminderLog;
+
+public class ReminderLogTest {
 
 	@BeforeClass
 	public static void setUpBeforeClass() throws Exception {
@@ -16,11 +18,11 @@ public class EventsLogTest {
 	public static void tearDownAfterClass() throws Exception {
 	}
 	
-	private static EventsLog test = new EventsLog("4/16/17", "This is a test");
-	private static EventsLog invalidTest;
+	private static ReminderLog test = new ReminderLog("4/16/17", "This is a test");
+	private static ReminderLog invalidTest;
 	
 	@Test
-	public void EventsLogTestConstructor() {
+	public void ReminderLogTestConstructor() {
 		assertNotNull(test);
 		assertNull(invalidTest);
 	}
